pyproject.toml
[tool.poetry]
name = "tap_treez"
version = "0.1.0"
description = ""
authors = ["Ryan Allen <rcallen89@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.25.1"
singer-python = "^5.9.1"
python-dotenv = "^0.15.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.1"
autopep8 = "^1.5.4"
pylint = "^2.6.0"
coverage = "^5.3.1"
pytest-cov = "^2.11.1"
vcrpy = "^4.1.1"
pytest-vcr = "^1.0.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"